pyproject.toml
[project]
name = "assignment-1"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "python-levenshtein>=0.27.3",
    "python-whois>=0.9.5",
    "streamlit>=1.53.1",
    "tldextract>=5.3.0",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
